started extracting stuff; failover test added
