+++ src/main/java/redis/clients/jedis/JedisPubSubAdapter.java
@@ -1,0 +1,30 @@
+package redis.clients.jedis;
+
+/**
+* Created by piotrturek on 29/03/15.
+*/
+class JedisPubSubAdapter extends JedisPubSub {
+    @Override
+    public void onMessage(String channel, String message) {
+    }
+
+    @Override
+    public void onPMessage(String pattern, String channel, String message) {
+    }
+
+    @Override
+    public void onPSubscribe(String pattern, int subscribedChannels) {
+    }
+
+    @Override
+    public void onPUnsubscribe(String pattern, int subscribedChannels) {
+    }
+
+    @Override
+    public void onSubscribe(String channel, int subscribedChannels) {
+    }
+
+    @Override
+    public void onUnsubscribe(String channel, int subscribedChannels) {
+    }
+}
